@@ -1,14 +1,15 @@
 {
     "name": "biberlabs/ddd-embeddables",
-    "description": "A collection of re-usable value objects ready to use as Embeddables in Doctrine",
+    "description": "A collection of re-usable value objects ready to use as Embeddable in Doctrine entities",
     "type": "library",
     "license": "MIT",
     "require": {
-        "php": "^5.6 || ^7.0",
-        "doctrine/orm": "^2.5"
+        "php": "^8.0",
+        "doctrine/orm": "^2.11",
+        "composer-runtime-api": "^2"
     },
     "require-dev": {
-        "phpunit/phpunit": "^5.5"
+        "phpunit/phpunit": "^9.5"
     },
     "suggest": {
         "ext-gmp": "Calculate without integer limits",
