@@ -1,17 +1,19 @@
 <?php
 /**
  * IP address value object to represent an internet protocol address.
- * 
+ *
  * @see https://en.wikipedia.org/wiki/IPv6_address
- * 
+ *
  * @author  M. Yilmaz SUSLU <yilmazsuslu@gmail.com>
  * @license MIT
  *
  * @since   Sep 2016
  */
+
 namespace DDD\Embeddable;
 
 use Doctrine\ORM\Mapping as ORM;
+use InvalidArgumentException;
 
 /**
  * @ORM\Embeddable
@@ -22,29 +24,25 @@ class IpAddress
      * ipv4 or v6 address
      *
      * @ORM\Column(type="string", length=45, nullable=true)
-     * 
-     * @var string
      */
-    private $address;
+    private ?string $address = null;
 
     /**
      * Constructor
-     * 
-     * @see https://secure.php.net/manual/en/filter.filters.flags.php
-     * 
-     * @throws InvalidArgumentException
      *
-     * @param string $addr IP address.
+     * @see https://secure.php.net/manual/en/filter.filters.flags.php
+     *
+     * @throws InvalidArgumentException
      */
-    public function __construct($addr = null)
+    public function __construct(string $addr = null)
     {
-        if (!$addr) {
-            return; // early..
+        if (! $addr) {
+            return;
         }
 
         $filtered = filter_var($addr, FILTER_VALIDATE_IP);
         if ($filtered === false) {
-            throw new \InvalidArgumentException('Given IP '.$addr.' is not a valid IP address');
+            throw new InvalidArgumentException('Given IP ' . $addr . ' is not a valid IP address');
         }
 
         $this->address = $filtered;
@@ -52,10 +50,8 @@ class IpAddress
 
     /**
      * String representation of ip address.
-     * 
-     * @return string
      */
-    public function __toString()
+    public function __toString(): string
     {
         return $this->address ?: '';
     }
