@@ -1,15 +1,16 @@
 <?php
 /**
  * IP range value object
- * 
+ *
  * Thanks to Ross Tuck for this great article:
  * http://rosstuck.com/persisting-value-objects-in-doctrine/
- * 
+ *
  * @author  M. Yilmaz SUSLU <yilmazsuslu@gmail.com>
  * @license MIT
  *
  * @since   Sep 2016
  */
+
 namespace DDD\Embeddable;
 
 use Doctrine\ORM\Mapping as ORM;
@@ -22,28 +23,18 @@ class IpRange implements JsonSerializable
 {
     /**
      * beginning IP address of the range
-     * 
+     *
      * @ORM\Embedded(class="DDD\Embeddable\IpAddress")
-     * 
-     * @var IpAddress
      */
-    protected $startIp;
+    protected ?IpAddress $startIp = null;
 
     /**
      * high IP address of the range
      *
      * @ORM\Embedded(class="DDD\Embeddable\IpAddress")
-     * 
-     * @var IpAddress
      */
-    protected $endIp;
+    protected ?IpAddress $endIp = null;
 
-    /**
-     * Constructor.
-     * 
-     * @param IpAddress $startIp
-     * @param IpAddress $endIp
-     */
     public function __construct(IpAddress $startIp = null, IpAddress $endIp = null)
     {
         $this->startIp = $startIp;
@@ -52,38 +43,27 @@ class IpRange implements JsonSerializable
 
     /**
      * Returns the low IP addresses of this range.
-     * 
-     * @return IpAddress
      */
-    public function getStartIp()
+    public function getStartIp() : ?IpAddress
     {
         return $this->startIp;
     }
 
     /**
      * Returns the high IP addresses of this range.
-     * 
-     * @return IpAddress
      */
-    public function getEndIp()
+    public function getEndIp() : ?IpAddress
     {
         return $this->endIp;
     }
 
     /**
      * Create a new range from CIDR notation.
-     * CIDR notation is a compact representation of an IP address(es)
-     * and its associated routing prefix.
-     * 
-     * @static
-     *
-     * @param string $cidr
-     * 
-     * @return self
+     * CIDR notation is a compact representation of an IP address(es) and its associated routing prefix.
      */
-    public static function createFromCidrNotation($cidr)
+    public static function fromCIDR(string $cidr) : IpRange
     {
-        list($subnet, $bits) = explode('/', $cidr);
+        [$subnet, $bits] = explode('/', $cidr);
         $start               = long2ip((ip2long($subnet)) & ((-1 << (32 - (int)$bits))));
         $end                 = long2ip((ip2long($subnet)) + pow(2, (32 - (int)$bits))-1);
 
@@ -92,49 +72,41 @@ class IpRange implements JsonSerializable
 
     /**
      * String representation of a range.
-     * 
+     *
      * Example output: "192.168.0.10 - 192.168.0.255"
-     *    
-     * @return string
      */
-    public function __toString()
+    public function __toString() : string
     {
-        return $this->isEmpty() ? '' : (string) $this->startIp.' - '. $this->endIp;
+        return $this->isEmpty() ? '' : $this->startIp . ' - ' . $this->endIp;
     }
 
     /**
      * Array representation of the ip range
-     * 
-     * @return array
      */
-    public function toArray()
+    public function toArray() : array
     {
         if ($this->isEmpty()) {
             return [];
         }
 
         return [
-            'startIp' => (string) $this->getStartIp(),
-            'endIp'   => (string) $this->getEndIp(),
+            'startIp' => (string)$this->getStartIp(),
+            'endIp' => (string)$this->getEndIp(),
         ];
     }
 
     /**
      * Returns boolean TRUE if the range is empty, false otherwise.
-     * 
-     * @return boolean
      */
-    public function isEmpty()
+    public function isEmpty() : bool
     {
         return $this->startIp === null || $this->endIp === null;
     }
 
     /**
      * Implement json serializable interface.
-     * 
-     * @return array
      */
-    public function jsonSerialize()
+    public function jsonSerialize() : array
     {
         return $this->toArray();
     }
