@@ -7,51 +7,66 @@
  *
  * @since   Sep 2016
  */
+
 namespace Test\Embeddable;
 
 use DDD\Embeddable\GeoPoint;
+use PHPUnit\Framework\TestCase;
 
-class GeoPointTest extends \PHPUnit_Framework_TestCase
+class GeoPointTest extends TestCase
 {
-    public function testValidGeoPoint()
+    public function testValidGeoPoint() : void
     {
-        $geoPoint = new GeoPoint(24, 34);
-        $this->assertEquals($geoPoint->toArray(), ['lat' => 24, 'lng' => 34]);
-        $this->assertEquals($geoPoint->jsonSerialize(), ['lat' => 24, 'lng' => 34]);
-        $this->assertEquals($geoPoint->toElastic(), ['lat' => 24, 'lon' => 34]);
+        $underTest = new GeoPoint(24, 34);
+        $this->assertSame(['lat' => 24.0, 'lng' => 34.0], $underTest->toArray());
+        $this->assertSame(['lat' => 24.0, 'lng' => 34.0], $underTest->jsonSerialize());
+        $this->assertSame(['lat' => 24.0, 'lon' => 34.0], $underTest->toElastic());
     }
 
-    /**
-     * @expectedException \InvalidArgumentException
-     */
-    public function testInvalidGeoPoint()
+    public function testInvalidGeoPoint() : void
     {
-        $geoPoint = new GeoPoint(-300, 300);
+        $this->expectException(\InvalidArgumentException::class);
+        new GeoPoint(-300, 300);
     }
 
-    public function testEmptyState()
+    public function testEmptyState() : void
     {
-        $p = new GeoPoint();
-        $this->assertInstanceOf(GeoPoint::class, $p);
-        $this->assertEquals([], $p->toArray());
-        $this->assertEquals([], $p->toElastic());
+        $underTest = new GeoPoint();
+        $this->assertInstanceOf(GeoPoint::class, $underTest);
+        $this->assertSame([], $underTest->toArray());
+        $this->assertSame([], $underTest->toElastic());
+        $this->assertSame('', (string)$underTest);
     }
 
-    public function testStringRepresentationAndEmptyState()
+    public function testStringRepresentation() : void
     {
-        $point = new GeoPoint(41.520112, 29.453401);
-        $this->assertEquals((string) $point, '41.520112 29.453401');
-
-        $point = new GeoPoint();
-        $this->assertEquals(null, (string) $point);
-        $this->assertInternalType('array', $point->toArray());
+        $underTest = new GeoPoint(41.520112, 29.453401);
+        $this->assertSame('41.520112 29.453401', (string)$underTest);
     }
 
-    public function testElasticsearchSupport()
+    public function testToElastic() : void
     {
-        $point   = new GeoPoint(41, 29);
-        $elastic = ['lat' => 41, 'lon' => 29];
-        $this->assertEquals($elastic, $point->toElastic());
-        $this->assertNotEquals($elastic, $point->toArray());
+        // Given
+        $underTest = new GeoPoint(41, 29);
+        $expected  = ['lat' => 41.0, 'lon' => 29.0];
+
+        // When
+        $actual = $underTest->toElastic();
+
+        // Then
+        $this->assertSame($expected, $actual);
+    }
+
+    public function testToArray() : void
+    {
+        // Given
+        $underTest = new GeoPoint(42, 30);
+        $expected  = ['lat' => 42.0, 'lng' => 30.0];
+
+        // When
+        $actual = $underTest->toArray();
+
+        // Then
+        $this->assertSame($expected, $actual);
     }
 }
